+++ components/errors/src/lib.rs
@@ -1,4 +1,4 @@
-#![allow(unused_doc_comment)]
+#![allow(unused_doc_comments)]
 
 #[macro_use]
 extern crate error_chain;